+++ modules/elementary_functions/builtin/cpp/Gateway.cpp
@@ -106,6 +106,7 @@
 #include "isvectorBuiltin.hpp"
 #include "isscalarBuiltin.hpp"
 #include "triuBuiltin.hpp"
+#include "trilBuiltin.hpp"
 //=============================================================================
 using namespace Nelson;
 //=============================================================================
@@ -284,7 +285,8 @@ static const nlsGateway gateway[] = {
     { "isscalar", (void*)Nelson::ElementaryFunctionsGateway::isscalarBuiltin, 1, 1, CPP_BUILTIN },
     { "triu", (void*)Nelson::ElementaryFunctionsGateway::triuBuiltin, 1, 2,
         CPP_BUILTIN_WITH_EVALUATOR },
-
+    { "tril", (void*)Nelson::ElementaryFunctionsGateway::trilBuiltin, 1, 2,
+        CPP_BUILTIN_WITH_EVALUATOR },
 };
 //=============================================================================
 NLSGATEWAYFUNC(gateway)